persist-txn: move StepForward to persist-types
Otherwise mz-repr will have to depend on persist-txn, which means that
persist-txn iteration would invalidate most of the codebase.
