+++ src/persist-types/src/lib.rs
@@ -167,3 +167,21 @@ pub trait Opaque: PartialEq + Clone + Sized + 'static {
     /// have yet been successful.
     fn initial() -> Self;
 }
+
+/// Advance a timestamp by the least amount possible such that
+/// `ts.less_than(ts.step_forward())` is true.
+///
+/// TODO(txn): Unify this with repr's TimestampManipulation. Or, ideally, get
+/// rid of it entirely by making persist-txn methods take an `advance_to`
+/// argument.
+pub trait StepForward {
+    /// Advance a timestamp by the least amount possible such that
+    /// `ts.less_than(ts.step_forward())` is true. Panic if unable to do so.
+    fn step_forward(&self) -> Self;
+}
+
+impl StepForward for u64 {
+    fn step_forward(&self) -> Self {
+        self.checked_add(1).unwrap()
+    }
+}
